ti test with parameters

+++ src/io/amal_base64.cpp
@@ -15,7 +15,7 @@ auto amal_base64 = [](const std::vector<std::string> &param) {
     auto maximum_literal_length = 65500 / line_width * line_width; // MSVC cannot deal with literal with length > 65535
     fmt::print(fo, "#include <taichi/common/util.h>\n\nTC_NAMESPACE_BEGIN\n\n\n");
     int num_literals = 0;
-    for (int l = 0; l < encoded.size(); l += maximum_literal_length) {
+    for (int l = 0; l < (int)encoded.size(); l += maximum_literal_length) {
       fmt::print(fo, "const std::string {}_{:04d} =\n", param[1], num_literals);
       num_literals += 1;
       for (int i = l; i < std::min((int)encoded.size(), l + maximum_literal_length); i += line_width) {
